Separate executable python scripts from modules
Move all importable code into a new package lib from src, leaving only
code that should be executed directly in src (ie, scripts that can be
run as __main__). This involves changing a bunch of imports and removing
a couple of `if __name__ == '__main__'` conditionals.

This commit is *refactoring only*, no functionality is changed.